requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-auth==2.28.1
django-allauth==0.61.1
dj-rest-auth==5.0.2
djangorestframework-simplejwt==5.3.1 